pyproject.toml
[project]
name = "github-credential-broker"
version = "0.1.0"
description = "GitHub Actions OIDC credential broker"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0,<1",
    "httpx>=0.27.0,<1",
    "pydantic>=2.8.0,<3",
    "pydantic-settings>=2.4.0,<3",
    "pyjwt[crypto]>=2.9.0,<3",
    "pyyaml>=6.0.2,<7",
    "uvicorn[standard]>=0.30.0,<1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0,<9",
    "pytest-asyncio>=0.24.0,<1",
    "ruff>=0.6.0,<1",
]

[project.scripts]
github-credential-broker = "github_credential_broker.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/github_credential_broker"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

